記事部分をリファクタリング
